### 2019-12-20 #1977 @whedon
[PRE REVIEW]: Hoki: Making BPASS accessible through Python
**Submitting author:** @heloises (<a href="http://orcid.org/0000-0002-0504-4323">H. F. Stevance</a>)
**Repository:** <a href="https://github.com/HeloiseS/hoki" target ="_blank">https://github.com/HeloiseS/hoki</a>
**Version:** v1.3.2b0
