### 2019-12-20 #1977 @arfon
:wave: @HeloiseS, thanks for your submission to JOSS! Does this accompany a paper in a AAS journal?

Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission? Please don't @-mention them here (i.e. refer to them as `arfon` rather than @arfon).

Also, feel free to suggest reviewers that you're aware of from the wider community who you think might be able to give this a good review.
