### 2019-12-22 #1977 @arfon
@KshitijAggarwal and @ygrange - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
