### 2019-12-22 #1977 @ygrange
Not this year, but surely happy to have a look at it when back in
January �

--
Y. Grange

> Op 22 dec. 2019 om 22:39 heeft Arfon Smith <notifications@github.com>
het volgende geschreven:
>
> ﻿
> @KshitijAggarwal and @ygrange - would you be willing to review this
submission for JOSS? We carry out our checklist-driven reviews here in
GitHub issues and follow these guidelines:
https://joss.readthedocs.io/en/latest/review_criteria.html
>
