### 2020-01-06 #1987 @ygrange
Just wanted to check off the first bunch of boxes this week and that leads to one comment:
- The visual on the repo README page calls itself HOKI beta. Not really sure what the JOSS policy is but I would expect it to be that the published release is not in beta state. Probably more a question to @arfon.
