### 2020-01-06 #1987 @arfon
> - The visual on the repo README page calls itself HOKI beta. Not really sure what the JOSS policy is but I would expect it to be that the published release is not in beta state. Probably more a question to @arfon.

I guess this kind of depends upon what _kind_ of a beta this is. Wasn't Gmail famously in beta for > 10 years 😁?

@HeloiseS - can you advise?
