### 2020-01-07 #1987 @HeloiseS
Hello !

The main reason it says beta is because user testing has been limited to a few people at a workshop and me so far and it hadn't been reviewed by other programmers. I also expect that significant features I may not have thought about will be requested by BPASS and hoki users in the upcoming months.

I'm happy to take it out of beta after this review :) It makes more sense to me to have the first not-pre-release  come with the JOSS paper (rather than before), even if the code base remains in a state of continuous development.

What do you think? @arfon @ygrange
