### 2020-01-07 #1987 @arfon
> I'm happy to take it out of beta after this review :) It makes more sense to me to have the first not-pre-release come with the JOSS paper (rather than before), even if the code base remains in a state of continuous development.

Yes, this sounds very reasonable. This happens semi-regularly here.
