### 2020-01-07 #1987 @ygrange
You are making good points! My personal opinion (please ignore anything I say if you disagree) would be here that little usage testing certainly makes a good case for beta status but it would be nice for you to have an idea when you think it is ok.
Adding functionality (even if it is considered significant features) could be part of the standard release process so I would not necessarily make it a reason to keep it beta.
BTW (mostly for completeness and probably stating the blatently obvious): the reason I think it matters is that having it in beta phase may stop new potential users from using it.
