### 2020-01-07 #1987 @HeloiseS
> BTW (mostly for completeness and probably stating the blatently obvious): the reason I think it matters is that having it in beta phase may stop new potential users from using it.

That makes a lot of sense. I think I'll take the beta off when this review is done. Also you make a good point as to where do I draw the line with feature additions being a part of a pre release or a full release - I don't think there will ever be one. This code base will HAVE to evolve with the field so you're right, that's more a matter of big releases that a beta status.

:) Anyway, the beta is coming off when hoki's been signed off on you folks, thanks for the convo.
