### 2020-01-07 ebbaf2b
CMDs now indexable and at_log_age() method added
