+++ .travis.yml
@@ -35,7 +35,7 @@ env:
         # overridden underneath. They are defined here in order to save having
         # to repeat them for all configurations.
         - PYTHON_VERSION=3.7
-        - NUMPY_VERSION=stable
+        - NUMPY_VERSION=1.17
         - ASTROPY_VERSION=stable
         - MAIN_CMD='python setup.py'
         - SETUP_CMD='test'
