+++ .travis.yml
@@ -109,9 +109,9 @@ matrix:
         # time.
 
         - os: linux
-          env: PYTHON_VERSION=3.6 NUMPY_VERSION=1.15
-        - os: linux
-          env: NUMPY_VERSION=1.16
+          env: PYTHON_VERSION=3.6 NUMPY_VERSION=1.16
+        #- os: linux
+        #  env: PYTHON_VERSION=3.5 NUMPY_VERSION=1.16
 
         # Do a PEP8 test with flake8
         #- os: linux
