### 2020-01-16 #1987 @KshitijAggarwal
Thanks for suggesting me to review this. Great work on hoki. It is really interesting and would be very useful for the community. I haven't yet tested the functionality of the code, will do that soon.
Here are a few comments from the rest:

* Regarding authorship: As the co-authors haven't contributed to the code development directly, it is hard to know their contribution to their project. I presume they must've contributed significantly through discussions, etc, esp because they developed and maintain the BPASS code. Is that so? Maybe @arfon can suggest how to evaluate this in this situation.
* Under **Motivation**, an IDL code has been mentioned which was used to create synthetic CMDs before hoki. Is there any relevant link to the code and/or citation that can be added here? Similarly, I would recommend citing any other code/publication which have implemented any pre-processing on BPASS data products or any subset of the functionality of hoki.
* The tutorial link given [here](https://heloises.github.io/hoki/CMDs.html ) (and all other notebooks/pages under **Recipes**) should be updated with the GitHub tutorial repo.
Some typos:
Under **Research** in the paper:
* ‘asbolutely’ -> ‘absolutely'
* ‘invovled’ -> ‘involved’

This is my first JOSS review. Feel free to correct me or ask if any of the above doesn't make sense.
