### 2020-01-18 #1987 @HeloiseS
The JOSS guideline say that contribution doesn't have to be in code to justify authorship. JJ and elizabeth are the BPASS developers. I have needed their input and will continue to do so as I further develop the code.
