### 2020-01-18 #1987 @arfon
Thanks for flagging this @KshitijAggarwal. The authorship checkbox is there to prompt a conversation about authorship and is usually used by reviewers to identify potential _missing_ authors.

> The JOSS guideline say that contribution doesn't have to be in code to justify authorship.

That's correct.

> I have needed their input and will continue to do so as I further develop the code - it seems only fair to me they also get a citation.

Right, it sounds like they are making a major intellectual contribution to this effort even if they're not directly contributing code to this package.
