### 2020-01-20 e946fe8
added reference to BPASS manual and fixed typos
