+++ docs/_sources/Transient_rates.ipynb.txt
@@ -7,7 +7,7 @@
     " Transient rates \n",
     "---\n",
     "\n",
-    "Download all the Jupyter notebooks from: https://github.com/HeloiseS/hoki/tree/master/tutorials"
+    "Download all the Jupyter notebooks from: https://github.com/HeloiseS/hoki_tutorials"
    ]
   },
   {
