### 2020-01-20 #1987 @HeloiseS
Thanks so much for your comments @KshitijAggarwal !

> * Under **Motivation**, an IDL code has been mentioned which was used to create synthetic CMDs before hoki. Is there any relevant link to the code and/or citation that can be added here?

=> I Added a link to the user manual that talks about the IDL code. There isn't much in the way of published work on this since it's just a script that's been released with the data.

> Similarly, I would recommend citing any other code/publication which have implemented any pre-processing on BPASS data products or any subset of the functionality of hoki.

BPASS data have always required pre-processing and they have been used for years. This description would encompass dozens and dozens of papers (if not more). I'm not sure it would be useful to the reader - that's why we usually cite the core BPASS papers like Eldridge 2008, 2017 and Stanway 2018. I hope that's okay!

> * The tutorial link given [here](https://heloises.github.io/hoki/CMDs.html) (and all other notebooks/pages under **Recipes**) should be updated with the GitHub tutorial repo.
=> Fixed the links! Thanks!

>   Some typos:
>   Under **Research** in the paper:
> * ‘asbolutely’ -> ‘absolutely'
> * ‘invovled’ -> ‘involved’

=> Fixed! Thanks for spotting them :smiley:
