### 2020-01-21 #1987 @KshitijAggarwal
Thanks for replying to my comments. I think it is okay to just cite the BPASS papers in this case.

I looked at the code and really like the heavy use of Pandas Dataframes. The documentation is also pretty good and useful (so was the rickrolling!!). One small thing is that I prefer using [logging](https://docs.python.org/2/library/logging.html) instead of `prints` as it provides a pretty flexible logging system. It is not at all necessary for this but it is something you can have a look at if you haven't already.

I followed the tutorial notebooks which are well made and easy to follow. I encountered two errors during those, and have created two issues corresponding to them. Also, it wasn't very convenient to get data from Google Drive (esp. for the remote machine), but I don't think that is on `hoki`. I guess BPASS must've always done it this way. Still, if possible, it would be preferred to host this huge database on a server, from where it can be downloaded using command line tools.

https://github.com/HeloiseS/hoki/issues/4#issue-552557761
https://github.com/HeloiseS/hoki_tutorials/issues/1#issue-552556436
