### 2020-01-21 HeloiseS/hoki#5 @KshitijAggarwal
Installing hoki from source
For the JOSS [review](https://github.com/openjournals/joss-reviews/issues/1987), I tried to install `hoki` from Github by `python setup.py install`. The installation finished with a few warnings/errors:
```
git submodule command failed unexpectedly:
error: pathspec 'astropy_helpers' did not match any file(s) known to git.
Downloading 'astropy-helpers'; run setup.py with the --offline option to force offline installation.
.
.
.
reading manifest template 'MANIFEST.in'
warning: no files found matching 'CHANGES.rst'
warning: no files found matching '*.pyx' under directory 'hoki'
warning: no files found matching '*.c' under directory 'hoki'
warning: no files found matching '*.pxd' under directory 'hoki'
warning: no files found matching '*' under directory 'cextern'
warning: no files found matching '*' under directory 'scripts'
no previously-included directories found matching 'build'
no previously-included directories found matching 'docs/_build'
no previously-included directories found matching 'docs/api'
warning: no files found matching 'astropy_helpers/README.rst'
warning: no files found matching 'astropy_helpers/CHANGES.rst'
warning: no files found matching 'astropy_helpers/LICENSE.rst'
warning: no files found matching '*' under directory 'astropy_helpers/licenses'
warning: no files found matching 'astropy_helpers/ah_bootstrap.py'
warning: no files found matching '*.py' under directory 'astropy_helpers/astropy_helpers'
warning: no files found matching '*.pyx' under directory 'astropy_helpers/astropy_helpers'
warning: no files found matching '*.c' under directory 'astropy_helpers/astropy_helpers'
warning: no files found matching '*.h' under directory 'astropy_helpers/astropy_helpers'
warning: no files found matching '*.rst' under directory 'astropy_helpers/astropy_helpers'
warning: no files found matching '*' under directory 'astropy_helpers/astropy_helpers.egg-info'
warning: no files found matching '*' under directory 'astropy_helpers/astropy_helpers/sphinx'
no previously-included directories found matching 'astropy_helpers/build'
no previously-included directories found matching 'astropy_helpers/astropy_helpers/tests'
warning: no previously-included files matching '*.pyc' found anywhere in distribution
warning: no previously-included files matching '*.o' found anywhere in distribution
.
.
.
```

Although the code passed the tests, and tutorial notebooks didn't give any errors, it might be useful to look into the above warnings.
