### 2020-01-21 #1987 @KshitijAggarwal
I think it would be useful to add instructions on installing `hoki` from source (i.e Github). I did a simple install using `setup.py`, and it gave a few warnings ([here](https://github.com/HeloiseS/hoki/issues/5#issue-552602850)). Though all the tests were passed, and I was able to run the tutorial notebooks as well, it may be useful to look into the warnings and check if they might impact any untested function in `hoki`.

I don't have any more comments on this. It was really great reviewing this package, it is really well done and I am sure that it would be very useful for the people using BPASS data products. Once the issue (https://github.com/HeloiseS/hoki/issues/4) is addressed, I will check the last time on my checklist.
