+++ astropy_helpers/astropy_helpers/setup_helpers.py
@@ -1,0 +1,785 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+This module contains a number of utilities for use during
+setup/build/packaging that are useful to astropy as a whole.
+"""
+
+from __future__ import absolute_import
+
+import collections
+import os
+import re
+import subprocess
+import sys
+import traceback
+import warnings
+from configparser import ConfigParser
+import builtins
+
+from distutils import log
+from distutils.errors import DistutilsOptionError, DistutilsModuleError
+from distutils.core import Extension
+from distutils.core import Command
+from distutils.command.sdist import sdist as DistutilsSdist
+
+from setuptools import setup as setuptools_setup
+from setuptools.config import read_configuration
+from setuptools import find_packages as _find_packages
+
+from .distutils_helpers import (add_command_option, get_compiler_option,
+                                get_dummy_distribution, get_distutils_build_option,
+                                get_distutils_build_or_install_option)
+from .version_helpers import get_pkg_version_module, generate_version_py
+from .utils import (walk_skip_hidden, import_file, extends_doc,
+                    resolve_name, AstropyDeprecationWarning)
+
+from .commands.build_ext import AstropyHelpersBuildExt
+from .commands.test import AstropyTest
+
+# These imports are not used in this module, but are included for backwards
+# compat with older versions of this module
+from .utils import get_numpy_include_path, write_if_different  # noqa
+
+__all__ = ['register_commands', 'get_package_info']
+
+_module_state = {'registered_commands': None,
+                 'have_sphinx': False,
+                 'package_cache': None,
+                 'exclude_packages': set(),
+                 'excludes_too_late': False}
+
+try:
+    import sphinx  # noqa
+    _module_state['have_sphinx'] = True
+except ValueError as e:
+    # This can occur deep in the bowels of Sphinx's imports by way of docutils
+    # and an occurrence of this bug: http://bugs.python.org/issue18378
+    # In this case sphinx is effectively unusable
+    if 'unknown locale' in e.args[0]:
+        log.warn(
+            "Possible misconfiguration of one of the environment variables "
+            "LC_ALL, LC_CTYPES, LANG, or LANGUAGE.  For an example of how to "
+            "configure your system's language environment on OSX see "
+            "http://blog.remibergsma.com/2012/07/10/"
+            "setting-locales-correctly-on-mac-osx-terminal-application/")
+except ImportError:
+    pass
+except SyntaxError:
+    # occurs if markupsafe is recent version, which doesn't support Python 3.2
+    pass
+
+
+def setup(**kwargs):
+    """
+    A wrapper around setuptools' setup() function that automatically sets up
+    custom commands, generates a version file, and customizes the setup process
+    via the ``setup_package.py`` files.
+    """
+
+    # DEPRECATED: store the package name in a built-in variable so it's easy
+    # to get from other parts of the setup infrastructure. We should phase this
+    # out in packages that use it - the cookiecutter template should now be
+    # able to put the right package name where needed.
+    conf = read_configuration('setup.cfg')
+    builtins._ASTROPY_PACKAGE_NAME_ = conf['metadata']['name']
+
+    # Create a dictionary with setup command overrides. Note that this gets
+    # information about the package (name and version) from the setup.cfg file.
+    cmdclass = register_commands()
+
+    # Freeze build information in version.py. Note that this gets information
+    # about the package (name and version) from the setup.cfg file.
+    version = generate_version_py()
+
+    # Get configuration information from all of the various subpackages.
+    # See the docstring for setup_helpers.update_package_files for more
+    # details.
+    package_info = get_package_info()
[…]
