+++ astropy_helpers/astropy_helpers/sphinx/conf.py
@@ -1,0 +1,5 @@
+import warnings
+
+from sphinx_astropy.conf import * 
+
+warnings.warn("Note that astropy_helpers.sphinx.conf is deprecated - use sphinx_astropy.conf instead")
