### 2020-01-22 HeloiseS/hoki#5 @HeloiseS
Thanks!

I added the astropy_helpers repo to the git repo this is waht caused this:
``` error: pathspec 'astropy_helpers' did not match any file(s) known to git.```

For the rest, it's just that astropy helpers checks for things that I don't have in my package, like Cython files (.pxy) or C files (.c), I also don't keep track of changes in the code using and CHANGE.rst file so it can't kind it (I log changes in patches and versions on the release board of github FYI).

If the tests pass, then I'm not too worried.

Thanks!
