### 2020-01-22 e941b70
fixed the tutorials - updating the docs
