+++ hoki/data/settings.yaml
@@ -1,2 +1,2 @@
 models_path: /home/fste075/BPASS_hoki_dev/bpass-v2.2-newmodels/
-outputs_path: /home/fste075/BPASS_hoki_dev/bpass_v2.2.1_imf135_300/ #this for dev purposes
+outputs_path: /home/fste075/BPASS_hoki_dev/bpass_v2.2.1_imf135_300/
