### 2020-01-22 HeloiseS/hoki#4 @HeloiseS
Thanks for spotting this!

The problem here was with the data I provided - the pickle files needed updating!

It should all work now.
