### 2020-01-22 #1987 @KshitijAggarwal
Thanks for responding to my comments and issues so quickly. I reinstalled the package, re-ran the notebooks, and everything works perfectly. I have one small suggestion which I forgot to add yesterday (this is not essential for review btw). As you have mentioned in the tutorial notebook, `make()` in `CMD` class takes quite a bit of time to run. It might be useful to add a progress bar (like [tqdm](https://tqdm.github.io/)) on the for loop which loops over each line of the input file in this function. This will not only show the progress of the process but will also provide an estimated time to completion of the function.

The review is complete from my end now.
