### 2020-01-22 #1987 @HeloiseS
Thanks!!! I added an "enhancement" issue to keep that in mind for the next major release :smiley:
