### 2020-01-23 #1987 @ygrange
Ok, there we go:

The paper is clearly written, readable and concise. The software is clearly documented and the jupyter notebook tutorials are a pretty good way to get a feeling of how to use the tool. In general I really like this submission. So I fully agree with @KshitijAggarwal that this was an enjoyable package to review and that even without any experience in using BPASS data I can imagine this increasing the user base of those models greatly!

Here are my comments.

It is not really mentioned in the README what BPASS is. I think adding a sentence of two to the README will also make Hoki accessible to a larger audience of observational astronomers.

Installation of the master version of the code goes pretty smooth, even on an empty OS. Neat!

When installing Hoki on one of my test systems, pip reported that the bdist_wheel command was needed. I suggest to add the pip package `wheel` to the setup dependency list (though it probably exists in pretty much any realistic environment).

Small point of advice: add to the README how the current development version can be installed (`pip3 install .` in the checked-out repo is probably the most fool-proof option here). (NB: I think a comparable comment has also been given by the other reviewer). Also my personal preference is to be told that a package has tests I could execute if I install from source. I leave it up to you to decide if you agree and put it in the README.

Would be useful to mention Hoki does not work on python2 (yes, I know... Still many systems have it by default)

The Tutorials:
I really like the way you set up separate jupyter notebooks so that people get to play with the interface. Way to go!

I would assue people following the tutorial may be as new to the field as I am so I will basically flag all things that are not fully clear to me. If some things need too much explanation and the target audience should know, feel free to ignore (or comment).

* CMD tutorial:
  - I'm not really sure where I could get a "BPASS input" file from so I have to skip the first bit of the tutorial. Maybe providing such a file in the data directory (or point out that it can be found somewhere in the google drive) could be useful.

The three issues that you fixed were al in my review too :). Quite well-fixed :)

I was wondering: would it be possible to generate a small BPASS data set for testing purposes? If this is possible, I would advice that you use something like binder to give people a very low-threshold way to play with the notebooks (through a badge in the README). Probably 45GB downloads are not going to be too much appreciated by the binder people... Surely such data would be not very scientifically interesting and maybe the plots may not be illustrative.

Paper:
My take on the paper is that it is targeted at observational astronomers who could use some simulated data to compare to their observation. So I'm trying to read it as such.

You refer to "a code written in IDL". I would at least state its name (if it is at all known). Maybe a citation if you want to be nice.

One of the features is "Standard BPASS Constants". I am not really sure what this is supposed to mean.

Is BPASS a tool that I could download and run, or is it a sumulation that is provided by the BPASS team? I think it may be useful to clarify in a sentence or so in the paper.

Another thing I noticed is that since the BPASS data is relatively big, users may want to run Hoki on a remote machine. On the other hand the data is in a Google drive, making it not very trivial to download using wget or so. So I would advice you to consider implementing a method to query/download the BPASS data.
