### 2020-01-23 #1987 @ygrange
Small issue I noticed when having another glance at the readthedocs: you may want to link the notebook tutorials somewhere in the intro.
