+++ README.rst
@@ -25,6 +25,11 @@ Hoki isn't only versatile, it also takes care of the nitty gritty pre-processing
 
 **Spend less time on coding and more time on the science!**
 
+What is BPASS?
+-----
+
+*The Binary Populations And Spectral Synthesis (BPASS) code simulates stellar populations and follows their evolution until death. Including binary evolution is crucial to correctly interpreting astronomical observations. The detailed follow-up of the stellar evolution within the code allows the retreival of important information such as supernova and gravitational wave event rates, giving us the ability to understand the properties of the stellar populations that are the precursors of these events.*
+
    
 Installing hoki
 ^^^^^^^^^^^
