+++ README.rst
@@ -16,7 +16,7 @@
     :target: http://www.astropy.org
     :alt: Powered by Astropy Badge
     
-BRIDGING THE GAP BETWEEN OBSERVATION AND THEORY
+Bridging the Gap Between Observation and Theory
 ------------------------------------------------
 
 In order to facilitate the application of **BPASS** to a wide range of scientific investigations, we have developped the tools necessary for observers to take full advantage of our models in an intuitive manner. 
@@ -25,10 +25,9 @@ Hoki isn't only versatile, it also takes care of the nitty gritty pre-processing
 
 **Spend less time on coding and more time on the science!**
 
-What is BPASS?
-^^^^^^^^^^
+.. note:: 
 
-*The Binary Populations And Spectral Synthesis (BPASS) code simulates stellar populations and follows their evolution until death. Including binary evolution is crucial to correctly interpreting astronomical observations. The detailed follow-up of the stellar evolution within the code allows the retreival of important information such as supernova and gravitational wave event rates, giving us the ability to understand the properties of the stellar populations that are the precursors of these events.*
+   *The Binary Populations And Spectral Synthesis (BPASS) code simulates stellar populations and follows their evolution until death. Including binary evolution is crucial to correctly interpreting astronomical observations. The detailed follow-up of the stellar evolution within the code allows the retreival of important information such as supernova and gravitational wave event rates, giving us the ability to understand the properties of the stellar populations that are the precursors of these events.*
 
    
 Installing hoki
