+++ README.rst
@@ -26,7 +26,7 @@ Hoki isn't only versatile, it also takes care of the nitty gritty pre-processing
 
 **Spend less time on coding and more time on the science!**
 
-* **WHATS BPASS?** *
+***WHAT IS BPASS?***
    *The Binary Populations And Spectral Synthesis (BPASS) code simulates stellar populations and follows their evolution until death. Including binary evolution is crucial to correctly interpreting astronomical observations. The detailed follow-up of the stellar evolution within the code allows the retreival of important information such as supernova and gravitational wave event rates, giving us the ability to understand the properties of the stellar populations that are the precursors of these events.*
 
    
