### 2020-01-24 72568d2
Update README.rst
