+++ README.rst
@@ -27,6 +27,7 @@ Hoki isn't only versatile, it also takes care of the nitty gritty pre-processing
 **Spend less time on coding and more time on the science!**
 
 ***WHAT IS BPASS?***
+
    *The Binary Populations And Spectral Synthesis (BPASS) code simulates stellar populations and follows their evolution until death. Including binary evolution is crucial to correctly interpreting astronomical observations. The detailed follow-up of the stellar evolution within the code allows the retreival of important information such as supernova and gravitational wave event rates, giving us the ability to understand the properties of the stellar populations that are the precursors of these events.*
 
 ----
@@ -57,11 +58,14 @@ This way it'll show you each test as they pass or FAIL. In the pip and github ve
 
 
 **Requirements**
+
 The following packages are required by `hoki`. If you pip install the stable version from pypi it will all be done automatically.
 
-`astropy`, `numpy`, `pandas`, `matplotlib`, `pyyaml`, `wheel`
+.. code-block:: none
+   `astropy`, `numpy`, `pandas`, `matplotlib`, `pyyaml`, `wheel`
 
 **Note**
+
 Python 2 is not supported
 
 ----
