+++ docs/_sources/CMDs.ipynb.txt
@@ -39,13 +39,13 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "### Getting the Stellar models\n",
+    "### Getting the Stellar models and input files\n",
     "\n",
     "Colour Magnitude Diagrams (CMDs) are created by reading in the [BPASS](https://bpass.auckland.ac.nz/9.html) stellar models listed in the `model_input` files that can be found in the BPASS output folders (e.g. *bpass_v2.2.1_imf135_300*). These stellar models are in a separate directory (because it is absolutely massive), so you will have to download it separately if you want to run the following cells or make your own CMDs.\n",
     "\n",
     "**NOTE: You will be able to run the cells in the section \"Loading a pickled CMD file\" even if you can't download the full set of stellar models **\n",
     "\n",
-    "The stellar models can be downloaded from [the google drive](https://drive.google.com/drive/folders/1BS2w9hpdaJeul6-YtZum--F4gxWIPYXl) (*bpass-v2.2-newmodels*).\n",
+    "The stellar models and input files can be downloaded from [the google drive](https://drive.google.com/drive/folders/1BS2w9hpdaJeul6-YtZum--F4gxWIPYXl) (*bpass-v2.2-newmodels* for the models and e.g. *bpass_v2.2.1_imf135_300* to get the required `model_input` files).\n",
     "Then you will have to change the path to the models in the `settings.yaml` file -- this can be done using the `set_models_path` function contained in the `hoki.load` module: "
    ]
   },
