+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name = hoki
 # version should be PEP440 compatible (https://www.python.org/dev/peps/pep-0440/)
-version = 1.3.2-beta
+version = 1.3.3
 author = H. F. Stevance
 author_email = hfstevance@gmail.com
 description = A Python interface to BPASS - Bridging the gap between observation and theory 
@@ -15,7 +15,7 @@ python_requires = ">=3.5"
 [options]
 # install_requires should be formatted as a semicolon-separated list, e.g.:
 # install_requires = astropy; scipy; matplotlib
-install_requires = astropy; numpy; pandas; matplotlib; pyyaml
+install_requires = astropy; numpy; pandas; matplotlib; pyyaml; wheel
 zip_safe = False
 use_2to3 = False
 
