### 2020-01-24 #1987 @HeloiseS
Thanks @ygrange ! Here's a quick response/summary:

Fixes
----

- I added the description I have of BPASS in the readthedocs to the README :heavy_check_mark:
- added wheel to the requirements, although when i created a python 3.7 env to reinstall from scratch to reproduce your error, conda installed wheel automatically. Better safe than sorry so I added it anyway! :heavy_check_mark:
- Added instruction to install the dev environment with pip (btw I didn't know you could do that so thanks!!! I usually clone haha) :heavy_check_mark:
- mention that python 2 is not supported :heavy_check_mark:
- added the following sentence in the **Motivation** section:
> These data products are made available to the community through the BPASS website [...]

So people can get to it and make it clear it's the data products that the community uses.
Is that okay :question:
- added a download link to the jupyter notebooks in the quick start of the readthedocs :heavy_check_mark:

Clarifications
----
- The IDL code has no name, no refs no license. It's just a script that was also made available. The lack of infrastructure is why hoki was born.

- BPASS data sets for testing? I've provided some data so the unit tests can run on other peoples machine (in the data folder of hoki). For the 45GB of stellar models though, it's not possible to restrict the amount of models that are given though, it just doesn't work that way :(
