### 2020-01-26 #1987 @ygrange
I think this clarifies the status of the data :).

My way of reproducting the wheel issue was to build python 3.8 from source and removing all packages (the cluster I use for testing didn't have a reliable python3 module so we had to create one for this cause; don't even ask :) ).

On the clarifications:
* Ugh, don't we love code without license, name and probably even a clear origin? I think keeping the text as is, is ok in this case.
* I feared that that would be the case. Not sure if Binder would dislike such big files, but it's worth investigating (I can probably fork the repo myself whenever the review ends).

I think there are still a few of my minor issues not addressed, but thanks for all things you already did!
