### 2020-01-26 #1987 @HeloiseS
@ygrange
> I think there are still a few of my minor issues not addressed, but thanks for all things you already did!

Is there anything else you'd like me to do? I'm happy to make further modifications to the paper or code!

I notice you've not marked the "statement of need" box as finished, is there anything I can do to clarify that need?
