### 2020-01-27 #1987 @ygrange
Whoops, sorry that was my mistake! Statement of need is (very) clear!

I think the main issues you didn't mention in your comments are the two on terminology:
1. in one of the notebooks a "BPASS input file" is mentioned and I didn't really know what that was supposed to mean. Maybe would  be good to clarify
2. "Standard BPASS constants" in the feature list (in the paper) is also a bit I didn't really get the meaning of.

That's really pretty much it.
