### 2020-01-28 #1987 @HeloiseS
My bad, I forgot to say I changed the phrasing of the CMD jupyter notebook, now the description of the input file location and and naming convention is given at the same time as that of the stellar models:

> The stellar models and input files can be downloaded from the google drive (bpass-v2.2-newmodels for the models and e.g. bpass_v2.2.1_imf135_300 to get the required model_input files).

The notebook also tells people where they are going to be using these input files in this sentence:
> In hoki you will be creating a CMD() object instantiated with a model of your choosing (a particular IMF and metallicity)- for this you need to provide the location of a BPASS input file.

As for BPASS constants, I'm not sure what to say, they are constants used in BPASS - I don't mean to be facetious that's literally what they are! Like, the arrays containing the time bins and the number of years per time bin, a dictionary for the big BPASS array in the stellar models... They're just constants that come from the way the BPASS simulations are run and they are used by hoki and BPASS users.  It's not the main feature for sure, but it's there so I thought I'd declare it?
