### 2020-01-28 #1987 @ygrange
1. Ah ok, makes sense. I think what confused me, to begin with, was that the BPASS input files are actually outputs of the model. So this went hand in hand with my misunderstanding of the way BPASS data is provided. Since this is clarified on both ends, even I now would get it :).

2. I wasn't really sure whether to interpret that as physical constants (things like Hubble constant but maybe also the more standard physical constants like G) or something else. Again, I take the role of a somewhat ignorant person (which isn't too hard to play for me) who has some observations and wants to use the simulations to compare with.

So they are constants output by the model? Or are those actually the model inputs? Ore are they even higher-level constants that basically do not depend on the specific model that was executed? They are internally used in Hoki, but also by users you say. What would a user use them for? The docstring  seems to suggest I shouldn't really bother too much as a standard user.

So I think adding something like ", for internal Hoki usage, or advanced users" may clarify to 'people like me' that this is not something they should necessarily want to dive into.

Does that make sense?
