### 2020-01-28 #1987 @HeloiseS
@ygrange yes that's a great idea!

They are not physical constants, they're things defined in BPASS - that's why  some BPASS users might use them, but you're right that applying the term advance BPASS user makes sense here because for most people it will be hidden by hoki.

I added this to to paper now! :D
