### 2020-01-28 #1987 @arfon
@HeloiseS - it looks like both @ygrange and @KshitijAggarwal have completed their reviews now ✨

At this point could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
