### 2020-01-28 #1987 @arfon
> @arfon, So the DOI should be a different DOI than the one I already have for Hoki? I'm a bit confused.

Does the DOI you already have include all of the changes that have resulted from this review?
