### 2020-01-28 #1987 @HeloiseS
Not yet! I will push a final version and update pypi and remove beta status (was waiting for final approval).

The name of the Zenodo folder tho will evolve after each release - so I can can call this release the same thing as the paper but it's going to change. Is that okay?

Here is a link to the current Zenodo folder (Iatest release was as I was submitting to JOSS): https://zenodo.org/record/3588063#.XjCuJHUzbmE
