### 2020-01-29 #1987 @arfon
Great, thanks @HeloiseS. Please could you make the following adjustments to the [Zenodo release](https://zenodo.org/record/3629869#.XjC2tnUzbmE), making sure that:
